all tasks done

+++ controllers/cards.js
@@ -9,8 +9,9 @@ const getCards = (req, res) => {
 };
 
 const deleteCard = (req, res) => {
+  console.log(req.params.cardId, req.user._id);
   if (validator.isMongoId(req.params.cardId)) {
-    Card.findOneAndRemove(req.params.cardId)
+    Card.findOneAndRemove({ _id: req.params.cardId, owner: req.user._id })
       .then((user) => res.send({ data: user }))
       .catch((err) => res.status(500).send({ message: `Произошла ошибка ${err}` }));
   } else res.status(404).send({ message: 'Нет карточки с таким id' });
@@ -18,11 +19,9 @@ const deleteCard = (req, res) => {
 
 const createCard = (req, res) => {
   const { name, link } = req.body;
-  if (validator.isMongoId(req.params.userId)) {
-    Card.create({ name, link, owner: req.user._id })
-      .then((card) => res.send({ data: card }))
-      .catch((err) => res.status(500).send({ message: `Произошла ошибка ${err}` }));
-  } else res.status(404).send({ message: 'Нет пользователя с таким id' });
+  Card.create({ name, link, owner: req.user._id })
+    .then((card) => { console.log(card); res.send({ data: card }); })
+    .catch((err) => res.status(500).send({ message: `Произошла ошибка ${err}` }));
 };
 
 const likeCard = (req, res) => {
